components/com.iot.SoilMoistureSensor/requirements.txt
paho-mqtt
python-dotenv
python-decouple
